update move store user info to log in component
